% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapond-package.R
\docType{package}
\name{datapond-package}
\alias{datapond}
\alias{datapond-package}
\title{Data Pond}
\description{
An offline-first Shiny application for documenting dataset metadata.
}
\author{
\strong{Maintainer}: Data Pond Contributors \email{data-pond@example.org}

Authors:
\itemize{
  \item Data Pond Contributors \email{data-pond@example.org}
}

}
\keyword{internal}
