% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_paths.R
\name{datapond_resource}
\alias{datapond_resource}
\title{Locate a bundled Data Pond resource}
\usage{
datapond_resource(...)
}
\arguments{
\item{...}{Path components below \code{inst/extdata}.}
}
\value{
A normalized path to the requested installed resource.
}
\description{
Locate a bundled Data Pond resource
}
\keyword{internal}
