% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_server.R
\name{db_browser_server}
\alias{db_browser_server}
\title{Shiny module server for db_browser()}
\usage{
db_browser_server(id, height = "500px")
}
\arguments{
\item{id}{Character of length 1, module ID (must match UI)}

\item{height}{Height of data preview tables}
}
\value{
A Shiny module server function
}
\description{
Use \code{\link[=db_browser_ui]{db_browser_ui()}} and \code{db_browser_server()}
to include the data lake browser as a Shiny module in your own app.
}
\examples{
\dontrun{
# In your Shiny app:
server <- function(input, output, session) {
  db_browser_server("browser1")
}
}
}
\seealso{
Other shiny: 
\code{\link{db_browser}()},
\code{\link{db_browser_ui}()}
}
\concept{shiny}
