% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_ui.R
\name{db_browser_ui}
\alias{db_browser_ui}
\title{Shiny module UI for db_browser()}
\usage{
db_browser_ui(id, height = "500px")
}
\arguments{
\item{id}{Character of length 1, module ID}

\item{height}{Height of data preview tables (default "500px")}
}
\value{
A Shiny UI element
}
\description{
Use \code{db_browser_ui()} and \code{\link[=db_browser_server]{db_browser_server()}}
to include the data lake browser as a Shiny module in your own app.
}
\examples{
\dontrun{
# In your Shiny app UI:
ui <- fluidPage(
  db_browser_ui("browser1")
)

# In your Shiny app server:
server <- function(input, output, session) {
  db_browser_server("browser1")
}
}
}
\seealso{
Other shiny: 
\code{\link{db_browser}()},
\code{\link{db_browser_server}()}
}
\concept{shiny}
