% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{db_drop_view}
\alias{db_drop_view}
\title{Drop a view from DuckLake}
\usage{
db_drop_view(schema = "main", view, if_exists = FALSE)
}
\arguments{
\item{schema}{Schema name (default "main")}

\item{view}{View name}

\item{if_exists}{If TRUE, don't error if view doesn't exist (default FALSE)}
}
\value{
Invisibly returns TRUE on success
}
\description{
Drop a view from DuckLake
}
